Major changes in structure.
- Added in tMap isDeleted, a vector of boolean in order to check if a
trapezoid has being deleted or not. In this way the trapezoid will not
be removed by tMap.trapezoids but just flagged. In this way the
complexity go down.

- Trapezoid neighbors are handled now with ids istead of using pointers
do to the reallocation of vector that cause the losing of neighbors to
trapezoids in the tMap

- Removed the recursive algorithm that search for the target leaf in the
Dag and substitute it. Now that is done by chaning the reference to the
pointer: Trapezoid->Node

- DrawableTrapezoidalMap now receives a vector with only the trapezoids
that are not flagged

- The visualization of QueryPoint is broken due to the last change in
this list.

- Upgraded the trapezoidalmap_dataset.cpp with the new version

- Methods that split the Trapezoid are now indepent from the cases